Startet på jobben med å samle steg-kall fra frontend til en controller og begynt å skrive om slik at de returnerer den utvidede behandlingen.
