+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/behandling/BehandlingController.kt
@@ -5,7 +5,6 @@ import no.nav.familie.ba.sak.common.RessursUtils.illegalState
 import no.nav.familie.ba.sak.common.RessursUtils.ok
 import no.nav.familie.ba.sak.config.TaskRepositoryWrapper
 import no.nav.familie.ba.sak.ekstern.restDomene.RestFagsak
-import no.nav.familie.ba.sak.ekstern.restDomene.RestTilbakekreving
 import no.nav.familie.ba.sak.ekstern.restDomene.RestUtvidetBehandling
 import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingKategori
 import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingResultat
@@ -26,7 +25,6 @@ import no.nav.familie.kontrakter.felles.Ressurs
 import no.nav.security.token.support.core.api.ProtectedWithClaims
 import org.springframework.http.MediaType
 import org.springframework.http.ResponseEntity
-import org.springframework.transaction.annotation.Transactional
 import org.springframework.validation.annotation.Validated
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
@@ -108,22 +106,6 @@ class BehandlingController(
         return ResponseEntity.ok(Ressurs.success(utvidetBehandlingService.lagRestUtvidetBehandling(behandlingId = behandlingId)))
     }
 
-    @PutMapping(path = ["{behandlingId}/henlegg"], produces = [MediaType.APPLICATION_JSON_VALUE])
-    fun henleggBehandlingOgSendBrev(
-        @PathVariable(name = "behandlingId") behandlingId: Long,
-        @RequestBody henleggInfo: RestHenleggBehandlingInfo
-    ): ResponseEntity<Ressurs<RestFagsak>> {
-        tilgangService.verifiserHarTilgangTilHandling(
-            minimumBehandlerRolle = BehandlerRolle.SAKSBEHANDLER,
-            handling = "henlegge behandling"
-        )
-
-        val behandling = behandlingsService.hent(behandlingId)
-        val response = stegService.håndterHenleggBehandling(behandling, henleggInfo)
-
-        return ResponseEntity.ok(fagsakService.hentRestFagsak(fagsakId = response.fagsak.id))
-    }
-
     @GetMapping(path = ["/{behandlingId}/simulering"])
     fun hentSimulering(
         @PathVariable @BehandlingstilgangConstraint
@@ -150,27 +132,6 @@ class BehandlingController(
         val restFagsak = fagsakService.hentRestFagsak(fagsakId = behandling.fagsak.id)
 
         return ResponseEntity.ok(restFagsak)
-    }
-
-    @Transactional
-    @PostMapping(path = ["/{behandlingId}/tilbakekreving"])
-    fun lagreTilbakekrevingOgGåVidereTilNesteSteg(
-        @PathVariable behandlingId: Long,
-        @RequestBody restTilbakekreving: RestTilbakekreving?
-    ): ResponseEntity<Ressurs<RestFagsak>> {
-
-        val behandling = behandlingsService.hent(behandlingId)
-        stegService.håndterVurderTilbakekreving(behandling, restTilbakekreving)
-
-        return ResponseEntity.ok(fagsakService.hentRestFagsak(fagsakId = behandling.fagsak.id))
-    }
-
-    @PostMapping(path = ["/{behandlingId}/steg/behandlingsresultat"])
-    fun utledBehandlingsresultat(@PathVariable behandlingId: Long): ResponseEntity<Ressurs<RestFagsak>> {
-        val behandling = behandlingsService.hent(behandlingId)
-        stegService.håndterBehandlingsresultat(behandling)
-
-        return ResponseEntity.ok(fagsakService.hentRestFagsak(fagsakId = behandling.fagsak.id))
     }
 }
